requirements.txt
# AI Engine
aider-chat>=0.50.0

# Web Framework
flask>=2.3.0
flask-socketio>=5.3.0
eventlet>=0.33.0

# GitHub API
PyGithub>=2.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
